Add database

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -9,7 +9,7 @@ import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.User;
-import ru.yandex.practicum.filmorate.storage.film.InMemoryFilmStorage;
+import ru.yandex.practicum.filmorate.dal.storage.film.InMemoryFilmStorage;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
@@ -99,12 +99,14 @@ public class FilmService {
         }
     }
 
-    public boolean isIdNull(Integer id) {
+    public boolean isIdNull(long id) {
         return id == 0;
     }
 
     public boolean isDateNull(LocalDate date) {
         return date == null;
     }
+
+
 }
